separated ldupes_error from ldupes.h
